refactoring WIP
