+++ NetWorld-Immersion/timer.h
@@ -1,0 +1,3 @@
+DWORD healStartTime = 0;
+DWORD healTime = 2000;
+DWORD currentTime = 0;
