+++ NetWorld-Immersion/utils.h
@@ -1,0 +1,4 @@
+int clamp(int x, int minX, int maxX)
+{
+    return max(min(maxX, x), minX);
+}
